+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 MumbleDJ Changelog
 ==================
 
+### January 11, 2015 -- `v2.8.10`
+* Created a new configuration value in the General section called PlayerCommand. This allows the user to change between "ffmpeg" and "avconv" for playing audio files.
+* Added check for valid PlayerCommand value. If the value is invalid the bot will default to `ffmpeg`.
+
 ### December 26, 2015 -- `v2.8.9`
 * Fixed an incorrect `!currentsong` message for songs within playlists.
 
@@ -15,7 +19,7 @@ MumbleDJ Changelog
 * Added argument to !listsongs command to specify how many songs to list (thanks [@nkhoit](https://github.com/nkhoit)).
 
 ### December 14, 2015 -- `v2.8.5`
-* Added !listsongs command (thanks [@nkhoit](https://github.com/nkhoit)). 
+* Added !listsongs command (thanks [@nkhoit](https://github.com/nkhoit)).
 
 ### December 7, 2015 -- `v2.8.4`
 * YouTube and SoundCloud API keys are now stored in the configuration file instead of environment variables. Existing installations with API keys in environment variables will automatically be migrated to the configuration file (thanks [@Gamah](https://github.com/Gamah)).